WINDUPRULE-276: moved embedded cache providers into own package
