+++ rules-java-ee/addon/src/main/java/org/jboss/windup/rules/apps/javaee/rules/cache/DiscoverEmbeddedCacheMemcachedLibraryRuleProvider.java
@@ -1,4 +1,4 @@
-package org.jboss.windup.rules.apps.javaee.rules;
+package org.jboss.windup.rules.apps.javaee.rules.cache;
 
 import org.jboss.windup.config.AbstractRuleProvider;
 import org.jboss.windup.config.GraphRewrite;
@@ -23,9 +23,8 @@ import org.ocpsoft.rewrite.config.ConfigurationBuilder;
 import org.ocpsoft.rewrite.context.EvaluationContext;
 
 @RuleMetadata(phase = InitialAnalysisPhase.class, perform = "Discover Java libraries embedded")
-public class DiscoverEmbeddedCacheJBossCacheLibraryRuleProvider extends AbstractRuleProvider
+public class DiscoverEmbeddedCacheMemcachedLibraryRuleProvider extends AbstractRuleProvider
 {
-
     @Override
     public Configuration getConfiguration(RuleLoaderContext context)
     {
@@ -34,7 +33,7 @@ public class DiscoverEmbeddedCacheJBossCacheLibraryRuleProvider extends Abstract
         return ConfigurationBuilder.begin()
                     .addRule()
                     .when(Query.fromType(JarArchiveModel.class)
-                                .withProperty(FileModel.FILE_NAME, QueryPropertyComparisonType.REGEX, ".*jbosscache.*\\.jar$"))
+                                .withProperty(FileModel.FILE_NAME, QueryPropertyComparisonType.REGEX, ".*memcached.*\\.jar$"))
                     .perform(
                                 new AbstractIterationOperation<JarArchiveModel>()
                                 {
@@ -44,15 +43,15 @@ public class DiscoverEmbeddedCacheJBossCacheLibraryRuleProvider extends Abstract
                                         ClassificationModel classificationModel = classificationService.attachClassification(event, context,
                                                     fileResourceModel,
                                                     IssueCategoryRegistry.CLOUD_MANDATORY,
-                                                    "Caching - JBoss cache embedded library",
-                                                    "The application embedds a JBoss cache library.  \n"
+                                                    "Caching - Memcached client embedded library",
+                                                    "The application embedds a Memcached client library.  \n"
                                                     + "\n"
                                                     + "Cloud readiness issue as potential state information that is not persisted to a backing service.");
                                         classificationModel.setEffort(5);
                                         GraphContext graphContext = event.getGraphContext();
 
                                         TechnologyTagService technologyTagService = new TechnologyTagService(event.getGraphContext());
-                                        technologyTagService.addTagToFileModel(fileResourceModel, "JBoss cache (embedded)",
+                                        technologyTagService.addTagToFileModel(fileResourceModel, "Memcached client (embedded)",
                                                     TechnologyTagLevel.INFORMATIONAL);
 
                                     }
