+++ rules-java-ee/addon/src/main/java/org/jboss/windup/rules/apps/javaee/rules/cache/DiscoverEmbeddedCacheOscacheLibraryRuleProvider.java
@@ -1,4 +1,4 @@
-package org.jboss.windup.rules.apps.javaee.rules;
+package org.jboss.windup.rules.apps.javaee.rules.cache;
 
 import org.jboss.windup.config.AbstractRuleProvider;
 import org.jboss.windup.config.GraphRewrite;
